+++ src/objects/client_proto/PaginatedProto.js
@@ -3,10 +3,11 @@ const axios = require('axios')
 const methods = require('../../utils/methods')
 
 /**
- * Get a chunk of this logged in users notifications
- * @param  {Object}  opts={}       optional parameters
- * @param  {Number}  opts.limit=25 Number of items to fetch
- * @return {Promise<Object>}         chunk of notifications with paging info
+ * Get a Chunk of the notifications for this client
+ * @param  {Object}  opts={}        Optional parameters
+ * @param  {Number}  opts.limit=25  Number of items to fetch
+ * @param  {Number}  opts.next=null Nextpage token
+ * @return {Promise<Object>}        Chunk of notifications with paging info
  */
 Client.prototype.notifications_paginated = async function(opts = {}) {
     let Notification = require('../Notification')
@@ -15,20 +16,47 @@ Client.prototype.notifications_paginated = async function(opts = {}) {
     let data = await methods.paginated_data(`${instance.api}/news/my`, {
         limit: opts.limit || instance.paginated_size,
         key: 'news',
-        prev: opts.prev,
         next: opts.next,
-        headers: instance.headers
+        headers: await instance.headers
     })
 
-    data.items = data.items.map((item) => new Notification(item, { client: instance }))
+    data.items = data.items
+        .map(item => new Notification(item, { client: instance }))
+
     return data
 }
 
 /**
- * Get a chunk of this logged in users chats
- * @param  {Object}  opts={}       optional parameters
- * @param  {Number}  opts.limit=25 Number of items to fetch
- * @return {Promise<Object>}         chunk of chats with paging info
+ * Get a Chunk of posts from the feed of a channel
+ * @param  {Object}  opts={}            Optional parameters
+ * @param  {Number}  opts.limit=25      Number of items to fetch
+ * @param  {Number}  opts.next=null     Nextpage token
+ * @param  {Number}  opts.channel       Channel to fetch posts from
+ * @return {Promise<Object>}            Chunk of chat memebrs with paging info
+ */
+Client.prototype.channel_feed_paginated = async function(opts = {}) {
+    let Post = require('../Post')
+    let instance = opts.instance || this
+
+    let data = await methods.paginated_data(`${instance.api}/channels/${opts.channel.id || opts.channel}/items`, {
+        limit: opts.limit || instance.paginated_size,
+        key: 'content',
+        next: opts.next,
+        headers: await instance.headers
+    })
+
+    data.items = data.items
+        .map((item) => new Post(item.id, { client: instance, data: item }))
+
+    return data
+}
+
+/**
+ * Get a Chunk of this logged in users chats
+ * @param  {Object}  opts={}        Optional parameters
+ * @param  {Number}  opts.limit=25  Number of items to fetch
+ * @param  {Number}  opts.next=null Nextpage token
+ * @return {Promise<Object>}        Chunk of chats with paging info
  */
 Client.prototype.chats_paginated = async function(opts = {}) {
     let Chat = require('../Chat')
@@ -53,18 +81,18 @@ Client.prototype.chats_paginated = async function(opts = {}) {
         headers: await instance.sendbird_headers
     })
 
-    let chats = response.data.channels.map(
-        it => new Chat(it.channel_url, { client: instance, data: it })
-    )
+    let chats = response.data.channels
+        .map(it => new Chat(it.channel_url, { client: instance, data: it }))
     return { items: chats, paging: { next: response.data.next } }
-};
+}
 
 /**
- * Get a chunk of the messages in a chat
- * @param  {Object}  opts={}       optional parameters
- * @param  {Number}  opts.limit=25 Number of items to fetch
- * @param  {Number}  opts.chat    Chat to fetch messages from
- * @return {Promise<Object>}       Chunk of messages with paging info
+ * Get a Chunk of the messages in a chat
+ * @param  {Object}  opts={}        Optional parameters
+ * @param  {Number}  opts.limit=25  Number of items to fetch
+ * @param  {Number}  opts.chat      Chat to fetch messages from
+ * @param  {Number}  opts.next=null Nextpage token
+ * @return {Promise<Object>}        Chunk of messages with paging info
  */
 Client.prototype.chat_messages_paginated = async function(opts = {}) {
     let Message = require('../Message')
@@ -97,19 +125,19 @@ Client.prototype.chat_messages_paginated = async function(opts = {}) {
 
     let index = response.data.messages.length - 1
     let next_id = index >= 0 ? response.data.messages[index].message_id : null
-    let messages = response.data.messages.map(
-        it => new Message(it.message_id, it.channel_url, { client: instance, data: it })
-    )
+    let messages = response.data.messages
+        .map(it => new Message(it.message_id, it.channel_url, { client: instance, data: it }))
 
     return { items: messages, paging: { prev: null, next: next_id } }
 }
 
 /**
- * Get a chunk of the members of a chat
- * @param  {Object}  opts={}       optional parameters
- * @param  {Number}  opts.limit=25 Number of items to fetch
- * @param  {Number}  opts.chat     Chat to fetch members from
- * @return {Promise<Object>}       Chunk of chat memebrs with paging info
+ * Get a Chunk of the members of a chat
+ * @param  {Object}  opts={}        Optional parameters
+ * @param  {Number}  opts.limit=25  Number of items to fetch
+ * @param  {Number}  opts.chat      Chat to fetch members from
+ * @param  {Number}  opts.next=null Nextpage token
+ * @return {Promise<Object>}        Chunk of chat memebrs with paging info
  */
 Client.prototype.chat_members_paginated = async function(opts = {}) {
     let ChatUser = require('../ChatUser')
@@ -130,36 +158,195 @@ Client.prototype.chat_members_paginated = async function(opts = {}) {
         headers: await instance.sendbird_headers
     })
 
-    let members = response.data.members.map(
-        it => new ChatUser(it.user_id, opts.chat, { client: instance, chat_data: it, data: it })
-    )
+    let members = response.data.members
+        .map(it => new ChatUser(it.user_id, opts.chat, { client: instance, chat_data: it, data: it }))
 
     return { items: members, paging: { prev: null, next: response.data.next } }
 
 }
 
 /**
- * Get a chunk of posts from the feed of a channel
- * @param  {Object}  opts={}            optional parameters
- * @param  {Number}  opts.limit=25      Number of items to fetch
- * @param  {Number}  opts.channel       Channel to fetch posts from
- * @return {Promise<Object>}            Chunk of chat memebrs with paging info
+ * Get a Chunk of the posts marked as read by this client
+ * @param  {Object}  opts={}        Optional parameters
+ * @param  {Number}  opts.limit=25  Number of items to fetch
+ * @param  {Number}  opts.next=null Nextpage token
+ * @return {Promise<Post>}          Chunk of posts with paging info
  */
-Client.prototype.channel_feed_paginated = async function(opts = {}) {
+Client.prototype.reads_paginated = async function(opts = {}) {
     let Post = require('../Post')
     let instance = opts.instance || this
 
-    let params = {
-        limit: opts.limit || instance.paginated_size
-    }
+    let data = await methods.paginated_data(`${instance.api}/feeds/reads`, {
+            limit: opts.limit || instance.paginated_size,
+            key: 'content',
+            next: opts.next,
+            headers: await instance.headers
+        })
+        .catch(e => console.log(e.response.data))
 
-    if (opts.next) {
-        params.next = opts.next
-    }
+    data.items = data.items
+        .map(item => new Post(item.id, { client: instance, data: item }))
 
-    let data = await methods.paginated_data(`${instance.api}/channels/${opts.channel.id || opts.channel}/items`)
-    data.items = data.items.map((item) => new Post(item.id, { client: instance, data: item }))
     return data
-};
+}
+
+/**
+ * Get a Chunk of posts from collective
+ * @param  {Object}  opts={}        Optional parameters
+ * @param  {Number}  opts.limit=25  Number of items to fetch
+ * @param  {Number}  opts.next=null Nextpage token
+ * @return {Promise<Post>}          Chunk of posts with paging info
+ */
+Client.prototype.collective_paginated = async function(opts = {}) {
+    let Post = require('../Post')
+    let instance = opts.instance || this
+
+    let data = await methods.paginated_data(`${instance.api}/feeds/collective`, {
+        method: 'post',
+        limit: opts.limit || instance.paginated_size,
+        key: 'content',
+        next: opts.next,
+        headers: await instance.headers
+    })
+
+    data.items = data.items
+        .map(item => new Post(item.id, { client: instance, data: item }))
+
+    return data
+}
+
+/**
+ * Get a Chunk of featured posts
+ * @param  {Object}  opts={}        Optional parameters
+ * @param  {Number}  opts.limit=25  Number of items to fetch
+ * @param  {Number}  opts.next=null Nextpage token
+ * @return {Promise<Post>}          Chunk of posts with paging info
+ */
+Client.prototype.features_paginated = async function(opts = {}) {
+    let Post = require('../Post')
+    let instance = opts.instance || this
+
+    let data = await methods.paginated_data(`${instance.api}/feeds/featured`, {
+        limit: opts.limit || instance.paginated_size,
+        key: 'content',
+        next: opts.next,
+        headers: await instance.headers
+    })
+
+    data.items = data.items
+        .map(item => new Post(item.id, { client: instance, data: item }))
+
+    return data
+}
+
+/**
+ * Get a Chunk of the weekly digests
+ * @param  {Object}  opts={}             Optional parameters
+ * @param  {Number}  opts.limit=25       Number of items to fetch
+ * @param  {Number}  opts.next=null      Nextpage token
+ * @param  {Boolean} opts.comments=false Get comment data from each digest?
+ * @param  {Boolean} opts.contents=false Get content data from each digest?
+ * @return {Promise<Post>}               Chunk of posts with paging info
+ */
+Client.prototype.digests_paginated = async function(opts = {}) {
+    let Digest = require('../Digest')
+    let instance = opts.instance || this
+
+    let params = {
+        comments: ~~(opts.comments || false),
+        contents: ~~(opts.contents || false)
+    }
+
+    let data = await methods.paginated_data(`${instance.api}/digest_groups`, {
+        limit: opts.limit || instance.paginated_size,
+        next: opts.next,
+        params: params,
+        headers: await instance.headers
+    })
+
+    data.items = data.items
+        .map(item => new Digest(item.id, { client: this, data: item }))
+
+    return data
+}
+
+/**
+ * Get a Chunk of search results for a tag
+ * @param  {Object}  opts={}        Optional parameters
+ * @param  {String}  opts.query     Search query
+ * @param  {Number}  opts.limit=25  Number of items to fetch
+ * @param  {Number}  opts.next=null Nextpage token
+ * @return {Promise<Post>}          Chunk of posts with paging info
+ */
+Client.prototype.search_tags_paginated = async function(opts = {}) {
+    let Post = require('../Post')
+    let instance = opts.instance || this
+
+    let data = await methods.paginated_data(`${instance.api}/search/content`, {
+            limit: opts.limit || instance.paginated_size,
+            key: 'content',
+            next: opts.next,
+            params: { q: opts.query },
+            headers: await instance.headers
+        })
+        .catch(error => console.log(error.response))
+
+    data.items = data.items
+        .map(item => new Post(item.id, { client: this, data: item }))
+
+    return data
+}
+
+/**
+ * Get a Chunk of search results for a user query
+ * @param  {Object}  opts={}        Optional parameters
+ * @param  {String}  opts.query     Search query
+ * @param  {Number}  opts.limit=25  Number of items to fetch
+ * @param  {Number}  opts.next=null Nextpage token
+ * @return {Promise<User>}          Chunk of posts with paging info
+ */
+Client.prototype.search_users_paginated = async function(opts = {}) {
+    let User = require('../User')
+    let instance = opts.instance || this
+
+    let data = await methods.paginated_data(`${instance.api}/search/users`, {
+        limit: opts.limit || instance.paginated_size,
+        key: 'users',
+        next: opts.next,
+        params: { q: opts.query },
+        headers: await instance.headers
+    })
+
+    data.items = data.items
+        .map(item => new User(item.id, { client: this, data: item }))
+
+    return data
+}
+
+/**
+ * Get a chunk of search results for a chat query
+ * @param  {Object}  opts={}        Optional parameters
+ * @param  {String}  opts.query     Search query
+ * @param  {Number}  opts.limit=25  Number of items to fetch
+ * @param  {Number}  opts.next=null Nextpage token
+ * @return {Promise<Chat>}          chunk of posts with paging info
+ */
+Client.prototype.search_chats_paginated = async function(opts = {}) {
+    let Chat = require('../Chat')
+    let instance = opts.instance || this
+
+    let data = await methods.paginated_data(`${instance.api}/search/chats/channels`, {
+        limit: opts.limit || instance.paginated_size,
+        key: 'channels',
+        next: opts.next,
+        params: { q: opts.query },
+        headers: await instance.headers
+    })
+
+    data.items = data.items
+        .map(item => new Chat(item.channel_url, { client: this, data: item }))
+
+    return data
+}
 
 module.exports = Client